+++ source/whalecar-service/src/main/java/com/whalecar/domain/WxWebCarPrice.java
@@ -1,18 +1,20 @@
 package com.whalecar.domain;
 
+import java.math.BigDecimal;
+
 /**
  * Created by wufei on 14/11/24.
  */
 public class WxWebCarPrice {
-    public Integer getWebPrice() {
+    public BigDecimal getWebPrice() {
         return webPrice;
     }
 
-    public void setWebPrice(Integer webPrice) {
+    public void setWebPrice(BigDecimal webPrice) {
         this.webPrice = webPrice;
     }
 
-    private Integer webPrice;
+    private BigDecimal webPrice;
 
 
 }
